requirements.txt
streamlit==1.42.2
pandas==2.2.3
plotly==5.24.1
SQLAlchemy==2.0.38
psycopg[binary]==3.2.5
python-dotenv==1.0.1
pytest==8.3.5
